+++ src/edu/ufp/inf/en/siu/User.java
@@ -11,11 +11,10 @@ public abstract class User extends Person {
   private String password;
   private Vehicle mobilityType = new Vehicle();
   private DataBase db;
-  private RedBlackBST<Integer, Poi> visitedPoi = new RedBlackBST<>();
+  private RedBlackBST<Long, Poi> visitedPoi = new RedBlackBST<>();
 
   public User(String name, String address, String idNumber, Date birth) {
     super(name, address, idNumber, birth);
-    
   }
 
   public String getEmail() {
@@ -50,11 +49,11 @@ public abstract class User extends Person {
     this.db = db;
   }
 
-  public RedBlackBST<Integer, Poi> getVisitedPoi() {
+  public RedBlackBST<Long, Poi> getVisitedPoi() {
     return this.visitedPoi;
   }
 
-  public void setVisitedPoi(RedBlackBST<Integer, Poi> visitedPoi) {
+  public void setVisitedPoi(RedBlackBST<Long, Poi> visitedPoi) {
     this.visitedPoi = visitedPoi;
   }
